+++ jpro-routing/core-test/src/test/scala/one/jpro/platform/routing/crawl/TestMemoryTester.scala
@@ -1,9 +1,8 @@
 package one.jpro.platform.routing.crawl
 
 import javafx.scene.control.Label
-import one.jpro.platform.routing.RouteUtils._
 import one.jpro.platform.routing.crawl.TestUtils.{Page1, Page2}
-import one.jpro.platform.routing.{Route, RouteNode}
+import one.jpro.platform.routing.{Response, Route, RouteNode}
 import org.junit.jupiter.api.Test
 import simplefx.cores.default.inFX
 import simplefx.util.Predef.intercept
@@ -14,9 +13,9 @@ class TestMemoryTester {
   def simpleTest(): Unit = {
     def app = new RouteNode(null) {
       setRoute(Route.empty()
-        .and(getView("/", r => new Page1))
-        .and(getView("/page2", r => new Page2))
-        .and(getView("/page4", r => new Page2)))
+        .and(Route.get("/", r => Response.view(new Page1)))
+        .and(Route.get("/page2", r => Response.view(new Page2)))
+        .and(Route.get("/page4", r => Response.view(new Page2))))
     }
     val result = AppCrawler.crawlApp("http://localhost", () => app)
     MemoryTester.testForLeaks(result, () => app)
@@ -27,9 +26,9 @@ class TestMemoryTester {
     val page2 = new Page2
     def app = new RouteNode(null) {
       setRoute(Route.empty()
-        .and(getView("/", r => new Page1))
-        .and(getView("/page2", r => page2))
-        .and(getView("/page4", r => new Page2)))
+        .and(Route.get("/", r => Response.view(new Page1)))
+        .and(Route.get("/page2", r => Response.view(page2)))
+        .and(Route.get("/page4", r => Response.view(new Page2))))
     }
     val result = AppCrawler.crawlApp("http://localhost", () => app)
     intercept[Throwable](MemoryTester.testForLeaks(result, () => app))
@@ -41,9 +40,9 @@ class TestMemoryTester {
 
     def app = new RouteNode(null) {
       setRoute(Route.empty()
-        .and(getView("/", r => new Page1))
-        .and(getView("/page2", r => viewFromNode(node2)))
-        .and(getView("/page4", r => new Page2)))
+        .and(Route.get("/", r => Response.view(new Page1)))
+        .and(Route.get("/page2", r => Response.node(node2)))
+        .and(Route.get("/page4", r => Response.view(new Page2))))
     }
 
     val result = AppCrawler.crawlApp("http://localhost", () => app)
@@ -54,9 +53,9 @@ class TestMemoryTester {
   def simpleFailingTest3(): Unit = {
     val app = inFX(new RouteNode(null) {
       setRoute(Route.empty()
-        .and(getView("/", r => new Page1))
-        .and(getView("/page2", r => new Page2))
-        .and(getView("/page4", r => new Page2)))
+        .and(Route.get("/", r => Response.view(new Page1)))
+        .and(Route.get("/page2", r => Response.view(new Page2)))
+        .and(Route.get("/page4", r => Response.view(new Page2))))
     })
     val result = AppCrawler.crawlApp("http://localhost", () => app)
     intercept[Throwable](MemoryTester.testForLeaks(result, () => app)) // fails because the webapp is not collectable