+++ jpro-routing/example/src/main/scala/example/scala/TestExtensions.scala
@@ -1,7 +1,7 @@
 package example.scala
 
 
-import one.jpro.platform.routing.RouteUtils.{get, getNode, redirect}
+import one.jpro.platform.routing.Route.{get, redirect}
 import com.jpro.webapi.WebAPI
 import simplefx.all._
 import simplefx.core._
@@ -10,7 +10,7 @@ import simplefx.experimental._
 import java.util.function.Supplier
 import one.jpro.platform.routing.extensions.linkheader.LinkHeaderFilter.Link
 import fr.brouillard.oss.cssfx.CSSFX
-import one.jpro.platform.routing.{Filters, LinkUtil, Redirect, Request, Route, RouteNode, RouteUtils}
+import one.jpro.platform.routing.{Filters, LinkUtil, Redirect, Request, Response, Route, RouteNode, RouteUtils}
 import one.jpro.platform.routing.extensions.linkheader.LinkHeaderFilter
 import one.jpro.platform.routing.filter.container.ContainerFactory
 import one.jpro.platform.routing.sessionmanager.SessionManager
@@ -21,8 +21,8 @@ class TestExtensionsApp(stage: Stage) extends RouteNode(stage) {
   setRoute(
     Route.empty() /* StartRoute? */
       .and(redirect("/", "/home"))
-      .and(getNode("/home", (r) => new Label("HOME")))
-      .and(getNode("/secret", (r) => new Label("SECRET")))
+      .and(get("/home", (r) => Response.node(new Label("HOME"))))
+      .and(get("/secret", (r) => Response.node(new Label("SECRET"))))
       .filter(LinkHeaderFilter.create(Link("HOME","/home"), Link("SECRET","/secret")))
   )