+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 0.2.9-SNAPSHOT (TBD)
+ * Moved Methods from RouteUtils to Route - for more consistent APIs.
+ * Added Documentation for jpro-routing
 
 ### 0.2.8 (December 19, 2023)